+++ app/src/main/java/com/finalproject/shelter/presentation/config/AppProperties.java
@@ -1,0 +1,13 @@
+package com.finalproject.shelter.presentation.config;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+@ConfigurationProperties("app")
+public class AppProperties {
+
+    private String host;
+}
