+++ app/src/main/java/com/finalproject/shelter/presentation/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.finalproject.shelter.presentation.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
