--- app/src/main/java/com/finalproject/shelter/presentationLayer/config/AppProperties.java
@@ -1,13 +1,0 @@
-package com.finalproject.shelter.presentationLayer.config;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Data
-@Component
-@ConfigurationProperties("app")
-public class AppProperties {
-
-    private String host;
-}
