--- app/src/main/java/com/finalproject/shelter/presentationLayer/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package com.finalproject.shelter.presentationLayer.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
