+++ app/src/test/java/com/finalproject/shelter/business/settings/SettingsControllerTest.java
@@ -2,7 +2,7 @@ package com.finalproject.shelter.business.settings;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.finalproject.shelter.WithAccount;
-import com.finalproject.shelter.presentationLayer.controller.view.profile.ProfileSettingController;
+import com.finalproject.shelter.presentation.controller.view.profile.ProfileSettingController;
 import com.finalproject.shelter.domain.model.entity.userDomain.Account;
 import com.finalproject.shelter.domain.repository.AccountRepository;
 import com.finalproject.shelter.business.service.account.AccountService;
